+++ src/main/java/org/folio/rest/impl/InitAPIs.java
@@ -10,18 +10,30 @@ import com.fasterxml.jackson.databind.SerializationConfig;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
+import io.vertx.core.DeploymentOptions;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
+import io.vertx.core.Promise;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.jackson.DatabindCodec;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
+
+import org.folio.verticles.DataImportConsumerVerticle;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.support.AbstractApplicationContext;
 
 /**
  * The class initializes vertx context adding spring context
  */
 public class InitAPIs implements InitAPI {
   private final Logger logger = LogManager.getLogger(InitAPIs.class);
+
+  @Value("${mod.invoice.kafka.DataImportConsumerVerticle.instancesNumber:1}")
+  private int dataImportConsumerVerticleNumber;
+
+  @Value("${dataimport.consumer.verticle.mandatory:false}")
+  private boolean isConsumerVerticleMandatory;
 
   @Override
   public void init(Vertx vertx, Context context, Handler<AsyncResult<Boolean>> resultHandler) {
@@ -36,7 +48,15 @@ public class InitAPIs implements InitAPI {
         DatabindCodec.prettyMapper().setConfig(deserializationConfig);
 
         SpringContextUtil.init(vertx, context, ApplicationConfig.class);
-        handler.complete();
+        SpringContextUtil.autowireDependencies(this, context);
+
+        deployDataImportConsumerVerticle(vertx).onComplete(ar -> {
+          if (ar.failed() && isConsumerVerticleMandatory) {
+            handler.fail(ar.cause());
+          } else {
+            handler.complete();
+          }
+        });
       },
       result -> {
         if (result.succeeded()) {
@@ -46,5 +66,19 @@ public class InitAPIs implements InitAPI {
           resultHandler.handle(Future.failedFuture(result.cause()));
         }
       });
+  }
+
+  private Future<String> deployDataImportConsumerVerticle(Vertx vertx) {
+    Promise<String> promise = Promise.promise();
+    AbstractApplicationContext springContext = vertx.getOrCreateContext().get("springContext");
+
+    DeploymentOptions deploymentOptions = new DeploymentOptions()
+      .setInstances(dataImportConsumerVerticleNumber)
+      .setWorker(true);
+    vertx.deployVerticle(() -> springContext.getBean(DataImportConsumerVerticle.class), deploymentOptions, promise);
+
+    return promise.future()
+      .onSuccess(ar -> logger.info("DataImportConsumerVerticle verticles was successfully started"))
+      .onFailure(e -> logger.error("DataImportConsumerVerticle verticles was not successfully started", e));
   }
 }